+++ internal/app/cli/houston.go
@@ -7,6 +7,7 @@ import (
 	"net/url"
 
 	"github.com/spf13/cobra"
+	"github.com/ternaryss/houston/internal/app/cron"
 	"github.com/ternaryss/houston/internal/app/db"
 	"github.com/ternaryss/houston/internal/app/handlers"
 	icmd "github.com/ternaryss/houston/internal/app/handlers/cmd"
@@ -27,10 +28,20 @@ var rootCmd = &cobra.Command{
 		usersStore := db.NewUsersStore(dbProvider.Db())
 		webAppsStore := db.NewWebAppsStore(dbProvider.Db())
 		subscribersStore := db.NewSubscribersStore(dbProvider.Db())
+		healthChecksStore := db.NewHealthChecksStore(dbProvider.Db())
 		errorsHandler := handlers.NewErrorsHandler()
 		dashboardHandler := handlers.NewDashboardHandler(webAppsStore)
 		webAppsHandler := handlers.NewWebAppsHandler(webAppsStore, subscribersStore)
+		healthChecksHandler := handlers.NewHealthChecksHandler(settings, webAppsStore, healthChecksStore)
 		usersHandler := handlers.NewUsersHandler(settings, usersStore)
+		retentionScheduler := cron.NewRetentionScheduler(settings, healthChecksHandler)
+		fiveMinutesInterval := cron.NewFiveMinutesIntervalScheduler(healthChecksHandler)
+		fifteenMinutesInterval := cron.NewFifteenMinutesIntervalScheduler(healthChecksHandler)
+		oneHourInterval := cron.NewOneHourIntervalScheduler(healthChecksHandler)
+		retentionScheduler.Run()
+		fiveMinutesInterval.Run()
+		fifteenMinutesInterval.Run()
+		oneHourInterval.Run()
 		server := web.NewServer(settings, errorsHandler, dashboardHandler, webAppsHandler, usersHandler)
 		server.Run()
 	},