+++ internal/app/db/web_apps.go
@@ -67,7 +67,7 @@ func (s *webAppsStore) DeleteByIdAndUserEmail(id, usr string) error {
 }
 
 func (s *webAppsStore) GetByFilter(ftr types.Filter, pag types.Pagination) ([]*types.WebApp, error) {
-	query := fmt.Sprintf(`SELECT ID, NAME, URL, STATUS, INTERVAL, USER_EMAIL USEREMAIL, CREATED_AT CREATEDAT, MODIFIED_AT
+	query := fmt.Sprintf(`SELECT ID, NAME, URL, STATUS, INTERVAL, USER_EMAIL USEREMAIL, HEALTHY, CREATED_AT CREATEDAT, MODIFIED_AT
 		FROM WEB_APPS
 		WHERE LOWER(USER_EMAIL) = LOWER($1)
 		ORDER BY %s LIMIT $2 OFFSET $3`, ftr.Sort)
@@ -92,6 +92,7 @@ func (s *webAppsStore) GetByFilter(ftr types.Filter, pag types.Pagination) ([]*t
 			&result.Status,
 			&result.Interval,
 			&result.UserEmail,
+			&result.Healthy,
 			&createdAt,
 			&modifiedAt,
 		); err != nil {
@@ -110,7 +111,7 @@ func (s *webAppsStore) GetByIdAndUserEmail(id, usr string) (*types.WebApp, error
 	var app types.WebApp
 	var createdAt int64
 	var modifiedAt int64
-	query := `SELECT ID, NAME, URL, STATUS, INTERVAL, USER_EMAIL, CREATED_AT, MODIFIED_AT FROM WEB_APPS
+	query := `SELECT ID, NAME, URL, STATUS, INTERVAL, USER_EMAIL, HEALTHY, CREATED_AT, MODIFIED_AT FROM WEB_APPS
 		WHERE ID = $1 AND LOWER(USER_EMAIL) = LOWER($2)`
 
 	if err := s.db.QueryRow(query, id, usr).Scan(
@@ -120,6 +121,7 @@ func (s *webAppsStore) GetByIdAndUserEmail(id, usr string) (*types.WebApp, error
 		&app.Status,
 		&app.Interval,
 		&app.UserEmail,
+		&app.Healthy,
 		&createdAt,
 		&modifiedAt,
 	); err != nil {
@@ -132,10 +134,51 @@ func (s *webAppsStore) GetByIdAndUserEmail(id, usr string) (*types.WebApp, error
 	return &app, nil
 }
 
+func (s *webAppsStore) GetByIntervalOrderByNameAsc(itv string) ([]*types.WebApp, error) {
+	query := `SELECT ID, NAME, URL, STATUS, INTERVAL, USER_EMAIL, HEALTHY, CREATED_AT, MODIFIED_AT
+		FROM WEB_APPS
+		WHERE INTERVAL = $1
+		ORDER BY NAME ASC`
+	rows, err := s.db.Query(query, itv)
+
+	if err != nil {
+		return nil, err
+	}
+
+	defer rows.Close()
+	var collection []*types.WebApp
+
+	for rows.Next() {
+		var result types.WebApp
+		var createdAt int64
+		var modifiedAt int64
+
+		if err := rows.Scan(
+			&result.Id,
+			&result.Name,
+			&result.Url,
+			&result.Status,
+			&result.Interval,
+			&result.UserEmail,
+			&result.Healthy,
+			&createdAt,
+			&modifiedAt,
+		); err != nil {
+			return nil, err
+		}
+
+		result.CreatedAt = time.Unix(createdAt, 0).UTC()
+		result.ModifiedAt = time.Unix(modifiedAt, 0).UTC()
+		collection = append(collection, &result)
+	}
+
+	return collection, nil
+}
+
 func (s *webAppsStore) Insert(wap *types.WebApp) (*types.WebApp, error) {
 	var id string
-	query := `INSERT INTO WEB_APPS (ID, NAME, URL, STATUS, INTERVAL, USER_EMAIL, CREATED_AT, MODIFIED_AT)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING ID`
+	query := `INSERT INTO WEB_APPS (ID, NAME, URL, STATUS, INTERVAL, USER_EMAIL, HEALTHY, CREATED_AT, MODIFIED_AT)
+		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9) RETURNING ID`
 
 	if err := s.db.QueryRow(
 		query,
@@ -145,6 +188,7 @@ func (s *webAppsStore) Insert(wap *types.WebApp) (*types.WebApp, error) {
 		wap.Status,
 		wap.Interval,
 		wap.UserEmail,
+		wap.Healthy,
 		wap.CreatedAt.Unix(),
 		wap.ModifiedAt.Unix(),
 	).Scan(&id); err != nil {
@@ -156,8 +200,8 @@ func (s *webAppsStore) Insert(wap *types.WebApp) (*types.WebApp, error) {
 }
 
 func (s *webAppsStore) Update(wap *types.WebApp) (*types.WebApp, error) {
-	query := `UPDATE WEB_APPS SET NAME = $1, URL = $2, STATUS = $3, INTERVAL = $3, USER_EMAIL = $4, MODIFIED_AT = $5
-		WHERE ID = $6 AND USER_EMAIL = $4`
+	query := `UPDATE WEB_APPS SET NAME = $1, URL = $2, STATUS = $3, INTERVAL = $4, USER_EMAIL = $5, HEALTHY = $6, MODIFIED_AT = $7
+		WHERE ID = $8 AND USER_EMAIL = $5`
 
 	if _, err := s.db.Exec(
 		query,
@@ -166,6 +210,7 @@ func (s *webAppsStore) Update(wap *types.WebApp) (*types.WebApp, error) {
 		wap.Status,
 		wap.Interval,
 		wap.UserEmail,
+		wap.Healthy,
 		time.Now().UTC().Unix(),
 		wap.Id,
 	); err != nil {