+++ internal/app/types/entities.go
@@ -6,6 +6,8 @@ const (
 	Interval5M  string = "5M"
 	Interval15M string = "15M"
 	Interval1H  string = "1H"
+	HealthyOk   string = "OK"
+	HealthyErr  string = "ERROR"
 )
 
 var Intervals = []string{Interval5M, Interval15M, Interval1H}
@@ -35,6 +37,7 @@ type WebApp struct {
 	Status     int
 	Interval   string
 	UserEmail  string
+	Healthy    string
 	CreatedAt  time.Time
 	ModifiedAt time.Time
 }
@@ -48,6 +51,7 @@ func NewWebApp(name, url, interval, userEmail string, status int) *WebApp {
 		Status:     status,
 		Interval:   interval,
 		UserEmail:  userEmail,
+		Healthy:    "",
 		CreatedAt:  now,
 		ModifiedAt: now,
 	}
@@ -67,6 +71,25 @@ func NewSubscriber(webAppId, email string) *Subscriber {
 	return &Subscriber{
 		WebAppId:   webAppId,
 		Email:      email,
+		CreatedAt:  now,
+		ModifiedAt: now,
+	}
+}
+
+type HealthCheck struct {
+	Id         int64
+	WebAppId   string
+	Status     int
+	CreatedAt  time.Time
+	ModifiedAt time.Time
+}
+
+func NewHealthCheck(webAppId string, status int) *HealthCheck {
+	now := time.Now().UTC()
+
+	return &HealthCheck{
+		WebAppId:   webAppId,
+		Status:     status,
 		CreatedAt:  now,
 		ModifiedAt: now,
 	}